adding swagger and a readme

+++ src/user/user.controller.ts
@@ -1,42 +1,43 @@
 import {
   Controller,
   Get,
-  Post,
   Body,
   Patch,
   Param,
   Delete,
+  Req,
+  ForbiddenException,
 } from "@nestjs/common";
 import { UserService } from "./user.service";
-import { CreateUserDto } from "./dto/create-user.dto";
 import { UpdateUserDto } from "./dto/update-user.dto";
+import { Request } from "express";
+import { ApiBearerAuth, ApiTags } from "@nestjs/swagger";
+import { User } from "./schemas/user.schema";
 
+@ApiBearerAuth()
+@ApiTags("user")
 @Controller("user")
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
-  @Post()
-  create(@Body() createUserDto: CreateUserDto) {
-    return this.userService.create(createUserDto);
-  }
-
-  @Get()
-  findAll() {
-    return this.userService.findAll();
-  }
-
   @Get(":id")
-  findOne(@Param("id") id: string) {
+  async findOne(@Param("id") id: string): Promise<User> {
     return this.userService.findOne(id);
   }
 
   @Patch(":id")
-  update(@Param("id") id: string, @Body() updateUserDto: UpdateUserDto) {
+  async update(
+    @Param("id") id: string,
+    @Body() updateUserDto: UpdateUserDto,
+    @Req() req: Request,
+  ): Promise<User> {
+    if (id !== req.user.id) throw new ForbiddenException();
     return this.userService.update(id, updateUserDto);
   }
 
   @Delete(":id")
-  remove(@Param("id") id: string) {
+  async remove(@Param("id") id: string, @Req() req: Request): Promise<User> {
+    if (id !== req.user.id) throw new ForbiddenException();
     return this.userService.remove(id);
   }
 }
